This fixes enable us to, sort of, "normally" test c connector
Summary:
Design doc or design overview:

**Docs impact**: none

Enabling testing for C Connector

Without these changes we're unable to properly run C connector
unittests.
One of the reasons is that after each test the driver sends COM_RESET
packet to which the engine returns an error. I suppose this is because
the engine does not properly handle COM_RESET packets or some other reason,
so that may be a task for the engine.
You may also encounter a problem with the engine crashing at some point.
I believe, that issue is already reported and tracked.
In the meantime, you can comment out the "test_bug28934" test in
ps_bugs.c to proceed with other tests.

UPD: double checked and there's indeed a missing logic on the engine side when the COM_RESET_CONNECTION packet is received.
Filed a task: https://memsql.atlassian.net/browse/DB-46851

Design doc or design overview:

**Docs impact**: none

Test Plan: the tests are running now

Reviewers: amakarovych-ua, iblinov-ua, sburlakov-ua

Reviewed By: amakarovych-ua

Subscribers: engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D43965