Explicitly conforming to C90
Summary:
Currently Visual Studio compiler on Windows conforms to C90 without the
option to specify another standard. That compels us to use C90 on all
platfroms until VS2019 supports newer standards.
Newer C is planned for the upcoming VS compiler versions.

Fixed our current tests to conform to C90.
GNU compiler on Linux is not strict about using C90, so it may allow
some of the things that wouldn't build by the strict VS2019 compiler.
But we'll have more confidence that the code we write will build on all
platforms.

Also fixed the wininstall/CMakelists.txt logic - it was missing one of
the plugin/auth package. If CMake is using VS
generator it works without the changes in the diff, but it fails for
other generators (e.g. NMake Makefiles) and it will be
better to include it and avoid potential build failures in the future.

Design doc or design overview:

**Docs impact**: none

Test Plan:
unit tests reran
build reran on Linux and Windows
for Windows it succeeded both with MSVC 16 2019 and "NMake Makefiles"

Reviewers: amakarovych-ua, iblinov-ua, sburlakov-ua, carl

Reviewed By: amakarovych-ua

Subscribers: engineering-list

JIRA Issues: PLAT-4922

Differential Revision: https://grizzly.internal.memcompute.com/D43802