Adding a test for the engine on the driver side
Summary:
This is related to the bug reported by HVR.
The fix was done on the engine side, so we're just making sure the
clients are getting the correct data. I believe, no current ciab images
contain this fix, so it's expected to fail now, but when testing against
the current master it succeeds, so once we update our testing cluster it
should succeed.

Design doc or design overview:

**Docs impact**: none

Test Plan: created a new test

Reviewers: amakarovych-ua, iblinov-ua, sburlakov-ua

Reviewed By: amakarovych-ua

Subscribers: engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D44049