Fixing the BIT extraction in the MySQL protocol
Summary:
Contrary to the rest of the BINARY data, our BIT is sent as a
big-endian, so when the client binds a small buffer (less than 8 bytes)
it may not be able to see the full value and will likely get 0 for a BIT
column (depending on the BIT data, of course).
We now parse BIT differently from the rest of the BINARY data, so when
it is fetched it looks exactly the same as the binary data. Therefore,
all the supported binary operations and conversion should be valid for
BIT as well.

It turns out also that MariaDB sends its binary in big-endian, whereas
we send it in the little-endian, so binding integer columns to the
binary data shows unexpected results. This is also fixed in this diff.

Also fixed some tests related to error handling for TIME structs.

Design doc or design overview:

**Docs impact**: none

Test Plan: ran tests, they passed

Reviewers: amakarovych-ua, iblinov-ua, sburlakov-ua

Reviewed By: amakarovych-ua

Subscribers: vtkachuk-ua, adam, engineering-list

JIRA Issues: PLAT-5028

Differential Revision: https://grizzly.internal.memcompute.com/D44643