Implementing the client-side prepared statements
Summary:
The feature is introduced with the NO_SSPS DSN flag which set to true by default. But in case this code has bugs (or it's just garbage) clients can override the value and use the binary protocol.

NOTE: all tests are now passing on Linux in Debug and RelWithDebInfo modes.

**Docs impact**:
After we have the ODBC docs ready, we should mention this flag.

Test Plan:
all the unit tests we have are now hitting this.
https://app.circleci.com/pipelines/github/memsql/singlestore-odbc-connector/347/workflows/7387fb55-1f58-45f5-8908-52d2887d50e3

Reviewers: carl, amakarovych-ua, sburlakov-ua, iblinov-ua, rob

Reviewed By: amakarovych-ua

Subscribers: engineering-list

JIRA Issues: PLAT-4814

Differential Revision: https://grizzly.internal.memcompute.com/D43566