Cleaning up after each test is run
Summary:
This diff truncates tables only from the odbc_test database which is
where the majority of our test data lives.

The tests are now running slower because of this. Besides, due to
running the same suite for both CSPS and SSPS, I'm increasing the
timeout to avoid annoying timeout errors which may occur when querying
the information_schema on older servers.

Design doc or design overview:

**Docs impact**: none

Test Plan:
ran tests
it is quite clear now that most of the errors are now related to the
actual test logic, rather than the node being down because of oom

there are some tests which create new databases - those are failing
because a newly created database consumes too much space

https://app.circleci.com/pipelines/github/memsql/singlestore-odbc-connector/393/workflows/b6da5825-082c-447d-8481-f3ea62d687d5

Reviewers: amakarovych-ua, iblinov-ua, sburlakov-ua

Reviewed By: amakarovych-ua

Subscribers: engineering-list

JIRA Issues: PLAT-5130

Differential Revision: https://grizzly.internal.memcompute.com/D44980