Leveraging Apd and Ipd types to insert params in CSPS
Summary:
Similar to what is implemented for the binary protocol, we implement
CHAR->BIT conversion and TIMESTAMP->DATE and TIMESTAMP->TIME
conversions.

Raising errors on invalid date/time data passed by the application.
Updated the tests that were previously contigent on the NoSsps flag to
be the mode-agnostic back again because the logic is now the same for
both CSPS and SSPS.

Design doc or design overview:

**Docs impact**: none

Test Plan:
added and updated tests

https://app.circleci.com/pipelines/github/memsql/singlestore-odbc-connector/397/workflows/8a2a39bb-e707-4714-8e9d-a028da41e62b

Reviewers: amakarovych-ua, iblinov-ua, sburlakov-ua

Reviewed By: amakarovych-ua

Subscribers: engineering-list

JIRA Issues: PLAT-5104

Differential Revision: https://grizzly.internal.memcompute.com/D44981