Chaged MADB_StmtTables to not ignore string lengths
Summary:
Changed t_bug14285620 test to expect SQL_success for Windows
It's DM passes a pointer to StringLength variable instead of NULL
As a result, we return SQL_SUCCESS instead of SWI

Changed sqlchar test to be run only for Unicode.
Windows DM replaces SQL_WCHAR to SQL_CHAR in the call to SQLGetData with ANSI driver.

**Design doc/spec**:
**Docs impact**: none

Test Plan:
https://app.circleci.com/pipelines/github/memsql/singlestore-odbc-connector/886/workflows/231ee8dc-3142-4ea8-98b2-5ab42ff830e2/jobs/2822
https://app.circleci.com/pipelines/github/memsql/singlestore-odbc-connector/887/workflows/8cf8a34e-9649-4a2c-a3cd-f1b13a98b5e2/jobs/2824

Reviewers: iblinov-ua, vtkachuk-ua, sburlakov-ua

Reviewed By: vtkachuk-ua

Subscribers: engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D46518