Added tests for SQLBulkOperations
Summary:
Added tests for SQLBulkOperations
Fixed precision issue in SQLBulkOperation.
When we are running SQLBukOperations, it creates an INPUT query under the hood.
When it tried to bind parameters for that input query, it used DisplaySize instead of a precision
**Design doc/spec**:
**Docs impact**: none

Test Plan:
Ran tests locally.
Planning to run them on CI

Reviewers: iblinov-ua, sburlakov-ua, stomash-ua, vtkachuk-ua, dPoltorak-ua

Reviewed By: iblinov-ua, dPoltorak-ua

Subscribers: engineering-list

JIRA Issues: PLAT-5417

Differential Revision: https://grizzly.internal.memcompute.com/D48715