Fixed Windows GUI
Summary:
NO_CACHE option was always enabled, because it was added as flag and as separate option.
TLS Peer Fingerprint option was made longer (it was too small to save the whole key).
Changed default cursor when changing it (before we changed defult cursor but only when initializing).

Deleted the following options:
 - Disable driver provided cursor support
 - Enable SQL_AUTO_IS_NULL
 - Pad Char to fullength with SPACE
 - Return “SQL_NULL_DATA” for zero date
 - Debug
 - Permit only specific TLS versions
 - Ignore schema in column specification
 - Server RSA public key

**Design doc/spec**:
**Docs impact**: none

Test Plan: manualy tested on Windows

Reviewers: iblinov-ua, sburlakov-ua, dPoltorak-ua, vtkachuk-ua

Reviewed By: dPoltorak-ua

Subscribers: engineering-list

JIRA Issues: PLAT-5614

Differential Revision: https://grizzly.internal.memcompute.com/D49649