Fixed driver for S2 v7.5
Summary:
Before we used `mysql_reset_connection` to clear resultset before getting a new one in CSPS mode.
This function was used incorrectly in the code and it was unsupported in older S2 versions.
Occurs that `mysql_reset_connection` don't do what is needed. It closes all open statements over a connection. We are reusing MYSQL_STMT object for CSPS (it violates some limitations of mariadb C client and should be refactored in the future). Because of it, we got crash.
I cleared the resultset manualy instead of using `mysql_reset_connection`.

Added 7.5 image to testing.

S2 7.5 started to support primary keys for columnstore.
There are some weord behaviour for this keys
https://memsql.atlassian.net/browse/DB-50840
so I changed queries to use rowstore.

In 7.5 column sizes for some types were changed.
I changed values that we are checking in the tests and filed a task to investigate this
https://memsql.atlassian.net/browse/PLAT-5724

**Design doc/spec**:
**Docs impact**: none

Test Plan: https://app.circleci.com/pipelines/github/memsql/singlestore-odbc-connector/1763/workflows/b59bf003-cd32-4285-948a-22e882a37b78

Reviewers: iblinov-ua, sburlakov-ua, vtkachuk-ua

Reviewed By: iblinov-ua, sburlakov-ua

Subscribers: engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D50269