Fixed type conversion during column binding
Summary: We ignored SQL_C_DEFAULT type handling and always treated it as MYSQL_TYPE_BLOB when retrieving the result. It cauese a problem when Excel Microsoft Query tries to get a smallint.

Test Plan:
Tested manually that it works with Excel Microsoft Query
Added a small test that uses SQL_C_DEFAULT
https://app.circleci.com/pipelines/github/memsql/singlestore-odbc-connector/1799/workflows/2b9feb54-7ae0-4d15-b4cc-4b194fbff0a3

Reviewers: iblinov-ua, vtkachuk-ua, sburlakov-ua

Reviewed By: sburlakov-ua

Subscribers: engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D50527