Fixed bug that a caused crash on calling SQLPrimaryKeys for information_schema
Summary:
We tried to get a table name from which to select primary keys.
Table names in information_schema are case-insensitive but we used SHOW TABLES LIKE to retrieve them and in this case comparison is case-sensitive.
As a result - table name was null.
Changed to:
 - use case-insansitive comparison for information_schema tables
 - return an empty result for non-existing tables

Test Plan: https://app.circleci.com/pipelines/github/memsql/singlestore-odbc-connector/2830

Reviewers: pmishchenko-ua

Reviewed By: pmishchenko-ua

Subscribers: engineering-list

JIRA Issues: PLAT-6416

Differential Revision: https://grizzly.internal.memcompute.com/D60950